pr fixes

+++ api/src/migration/common/1701368547549-add-soft-delete-for-relations.ts
@@ -14,11 +14,23 @@ export class AddSoftDeleteForRelations1701368547549 implements MigrationInterfac
         try {
             await queryRunner.query(
                 `
+    IF EXISTS (Select *
+               from INFORMATION_SCHEMA.TABLES
+               where TABLE_NAME = 'specification_details_LIB_Survey_CertificateAuthority'
+                 AND TABLE_SCHEMA = 'dry_dock')
+    BEGIN
                 alter table dry_dock.specification_details_LIB_Survey_CertificateAuthority
                 add active_status bit default 1 not null;
+    END
 
+    IF EXISTS (Select *
+               from INFORMATION_SCHEMA.TABLES
+               where TABLE_NAME = 'specification_details_j3_pms_agg_job'
+                 AND TABLE_SCHEMA = 'dry_dock')
+    BEGIN
                 alter table dry_dock.specification_details_j3_pms_agg_job
                 add active_status bit default 1 not null;
+    END
             `,
             );
 
@@ -44,11 +56,25 @@ export class AddSoftDeleteForRelations1701368547549 implements MigrationInterfac
     public async down(queryRunner: QueryRunner): Promise<void> {
         try {
             await queryRunner.query(`
+
+IF EXISTS (Select *
+               from INFORMATION_SCHEMA.TABLES
+               where TABLE_NAME = 'specification_details_LIB_Survey_CertificateAuthority'
+                 AND TABLE_SCHEMA = 'dry_dock')
+    BEGIN
             ALTER TABLE [dry_dock].[specification_details_LIB_Survey_CertificateAuthority]
             DROP COLUMN [active_status];
+    END
 
+
+    IF EXISTS (Select *
+               from INFORMATION_SCHEMA.TABLES
+               where TABLE_NAME = 'specification_details_j3_pms_agg_job'
+                 AND TABLE_SCHEMA = 'dry_dock')
+    BEGIN
             ALTER TABLE [dry_dock].[specification_details_j3_pms_agg_job]
             DROP COLUMN [active_status];
+    END
         `);
 
             await MigrationUtilsService.migrationLog(
